inline abstract class
